refactor: extract lightweight widget updater handles
Separates widget/activity update handles into widgetUpdaters.ts so
the main app bundle only imports lightweight update proxies, not the
full SwiftUI rendering components.
